fix: set tags for each size
