+++ packages/design-system/src/components/typography/headings/headings.module.scss
@@ -1,0 +1,46 @@
+@use "../../../base/" as *;
+@use "../../../functions/" as *;
+@use "../../../generator/" as *;
+@use "../../../mixins/" as *;
+@use "../text//text.module";
+
+@use "sass:map";
+
+.text {
+  @include text-ellipsis(1);
+  @include word-wrap;
+  font-size: font-size(text);
+}
+
+.headline {
+  font-size: font-size(small);
+}
+
+.default {
+  font-size: font-size(default);
+}
+
+.h0 {
+  font-size: font-size(h0);
+}
+
+.h1 {
+  font-size: font-size(h1);
+}
+
+.h2 {
+  font-size: font-size(h2);
+}
+.h3 {
+  font-size: font-size(h3);
+}
+.h4 {
+  font-size: font-size(h4);
+}
+
+.small {
+  font-size: font-size(small);
+}
+.tiny {
+  font-size: font-size(tiny);
+}
